+++ src/main/java/com/ndashimye/firstapp/user/UserRepository.java
@@ -6,12 +6,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
 
     @Query("select u from User AS u where u.profile.emailAddress = ?1")
     Optional<User> findUserByEmailAddress(String emailAddress);
-    boolean existsByUserId(Integer userId);
+    boolean existsByUserId(Long userId);
     boolean existsByUsername(String username);
 }
