+++ src/main/java/com/ndashimye/firstapp/userprofile/UserProfileRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
 
     boolean existsByEmailAddress(String emailAddress);
 }
