+++ src/main/java/com/ndashimye/firstapp/usersettings/UserSettingsRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserSettingsRepository extends JpaRepository<UserSettings, Integer> {
+public interface UserSettingsRepository extends JpaRepository<UserSettings, Long> {
 }
